+++ frontend/src/components/MessageBottomSheet.vue
